README.md
This is a very popular approach for affiliate marketers because it attracts visitors with high buyer intent.

Writing an in-depth review about the affiliate product you’re promoting is an effective marketing strategy.

Include the pros and the cons. While it might be tempting to just focus on the positives, don’t forget to mention the negatives as well. Readers want all of the information before they decide whether or not to purchase a product.