CONTRIBUTING.md
Imagine if you always knew the exact words to say to get people to buy products through your affiliate link. You’d quickly increase your affiliate sales, right?

Today we are going to be talking about the science and the art of persuasive marketing, so you know exactly what you need to say to sell more products with less effort.

Sure, you can get all the traffic in the world and have the best product in the world or the biggest email list, but that doesn’t matter if you don’t know how to persuade people to buy the products.